requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1

sqlalchemy==2.0.25
pydantic==2.6.1
pydantic-settings==2.1.0

python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

python-multipart==0.0.9

pytz==2024.1
